Reference GoogleService-Info.plist file in Xcode
This file contains all of the information required by the Firebase iOS
SDK to connect to our Firebase project. We don't want to actually
include it in our repository because we are an open source project and
expect folks who fork the project to supply their own plist file.
However, Xcode still needs to know about the file.

The expected way for this file to be created is by constructing it via a
shell script and supplying the correct Firebase credentials using
environment variables. This commit demonstrates how this can work within
the project's appcenter-post-clone.sh script.

Create GoogleService-Info.plist file in App Center