EventsMap: display notification on receiving new events
* Requests notification permission from App.js with default iOS dialog
for now. In the future, we should follow best practice and request
permission closer to where we need it and give the user context for
why we're asking.
* New events are determined by comparing old props and new props in
EventsMap.js. This means that when the app is restarted, we won't
remember what events we've already fetched. Future commits will improve
this by stashing events we've fetched already in local storage.
* Displays different notification depending on if there is one new event
or multiple